Big refactoring of request statistics code
Out of RequestStats, factored out a StatsEntry class which represents a single stats entry (URL). Previously the RequestStats class was actually doing two different things: On the global class level, it was holding track of the aggregated stats from all requests, and on the instance level holding the stats for single stats entries.

Now RequestStats should be instantiated and holds the global stats, as well as a dict of StatsEntry instances which holds the stats for single stats entries (URLs)