Fix an issue when trying to spawn 0 locusts
This was causing an issue when running distributed and trying to spawn fewer number of locusts than the number of slaves.

Also added two more tests that verifies the number of locusts we try to spawn when having uneven number of slaves connected (compared to the number of locusts we try to spawn).