Added the ability for locusts to be assigned a unique ID based on their type. I.e. If you run with two Locust types (called LocustA and LocustB) with equal weightings and 20 users, you'll have IDs 0 to 9 assigned to each LocustA and LocustB instance. Likewise, this is tracked across slaves, so if you ran with two slaves, slave1 would get ids 0 through 4 and slave2 would get ids 5 through 9.
There is additional handling added for increasing and decreasing IDs dynamically so changing the user_count is handled gracefully.

Example, if I run with the example above (2 slaves, 2 locust types and 20 users), and I increase the user count to 40, ids 0-4, 10-14 would be assigned to slave1, and
ids 5-9, 15-19 would be assigned to slave2. If users are decreased, the locusts with the highest ids are now killed first.
I.e. if I had the 20 users and decreased to 16 users, slave1 would have ids 0-3 and slave2 would have ids 5-8