Update dockerfile to slim down created image and tweak behaviors.
Change base image from py3.6 to py3.7-alpine. This shrinks the image size by about 600MB so that's nice.
Addition of the apk call is required for pip to actually build and of the dependencies for Locust.
Removal of docker_start portion is based on comments by mbeacom. Technically this makes the base image unusable without a consumer
adding their own Dockerfile with a COPY/ADD call or another ENTRYPOINT.