Add required --host option to examples
While the examples in the page focus on the master/slave concept, not providing the --host option will result in failure to start the test. It should be added, so that any user that copies the command from the documentation does not have to decipher an error message before getting the test to run.