Start web_ui later to avoid race adding UI routes
Flask throws an exception `AssertionError: A setup function was called after the first request was handled…` if it serves a request before all the routes are added, creating a race condition. Starting the web UI after firing `init` event avoids this race.

Requires `web_ui.start()` to be called manually when `create_web_ui()` or `WebUI()` are called or instantiated manually.