Introduce `LOCUST_SPAWN_AT_LEAST_ONE_USER_OF_EACH_USER_CLASS`
This boolean environment variable allows to change the behaviour of the users distribution. If `true`, it behaves as before by attempting to spawn at least one user of each class when the user count >= to the number of user classes.

If `false`, the real distribution is calculated, meaning that users with small weights won't always be spawned when a lower user count is set.

The test helper `patch_env` has been relocated in `test/util.py`.