Stop declaring "fake" class level variables in Environment, User and stats. Do it in constructor/instance level, where they belong instead.
The only benefit to having them on class level was so that they got documented properly, but it works as long as they are definied in __init__ as well.

While this shouldnt be a breaking change for any "normal" usage of Locust, I think we need to do some testing to ensure that we dont need to bump major version.

Added some type hints as well.